+++ src/NCronJob/Registry/JobRegistry.cs
@@ -145,7 +145,7 @@ internal sealed class JobRegistry
 
         public bool Equals(JobDefinition? x, JobDefinition? y) =>
             (x is null && y is null) || (x is not null && y is not null
-                                         && x.Type == y.Type && x.IsTypedJob
+                                         && x.JobFullName == y.JobFullName
                                          && x.Parameter == y.Parameter
                                          && x.CronExpression == y.CronExpression
                                          && x.TimeZone == y.TimeZone
@@ -153,7 +153,7 @@ internal sealed class JobRegistry
                                          && x.IsStartupJob == y.IsStartupJob);
 
         public int GetHashCode(JobDefinition obj) => HashCode.Combine(
-            obj.Type,
+            obj.JobFullName,
             obj.Parameter,
             obj.CronExpression,
             obj.TimeZone,