+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to **NCronJob** will be documented in this file. The project
 
 ## [Unreleased]
 
+- Allow the registration of the same untyped job with different names. Reported by [@Mangatome](https://github.com/Mangatome) in [#~253](https://github.com/NCronJob-Dev/NCronJob/issues/253). Fixed by [@nulltoken](https://github.com/nulltoken) in [#252](https://github.com/NCronJob-Dev/NCronJob/pull/252) and [#254](https://github.com/NCronJob-Dev/NCronJob/pull/254).
+
 ## [v4.4.0] - 2025-03-10
 
 ### Added